Update CentroidByColor
Code for the robotic lacrosse goalie. Filters by specified HSV value, erodes and dilates the mask, and overlays onto original frame. Also computes center of mass, and denotes which section (1-8) the center of mass is located. Chose to calculate section over precise angle for speed. Eventually this value will be sent to a motor over either ROS or through a Raspberry Pi.