Down, left, up, or right?
This code can read which direction you move your hand in front of a webcam. Accuracy is maybe 70%, but the algorithm is pretty slow (aka about a second to read each movement). Would like to add hand-detection (by shape of the hand) and color threshold, but obviously that is pretty region-dependent.